requirements.txt
camelot-py==0.10.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
ghostscript==0.7
pandas==1.3.5
PyPDF2==1.26.0
